Create `ProxyForm` template component
* Update `App.tsx`: new  **CreateProxy** &  **EditProxy** locations (src/renderer/App.tsx);
* Move & Update `CreateProxy` (src/renderer/components/dialogs/CreateProxy/CreateProxy.tsx, src/renderer/components/dialogs/CreateProxy/index.ts, src/renderer/components/dialogs/index.ts);
* Move & Update `EditProxy` (src/renderer/components/dialogs/EditProxy/EditProxy.tsx, src/renderer/components/dialogs/EditProxy/index.ts, src/renderer/components/dialogs/index.ts);
* Create `ProxyForm` template component (src/renderer/components/templates/ProxyForm/ProxyForm.tsx, src/renderer/components/templates/ProxyForm/index.ts, src/renderer/components/templates/index.ts);
* Update types for `proxy` (src/types/proxy.ts).
