+++ src/renderer/components/dialogs/index.ts
@@ -1,0 +1,2 @@
+export { default as CreateProxy } from "./CreateProxy";
+export { default as EditProxy } from "./EditProxy";
