+++ src/renderer/components/templates/ProxyForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProxyForm";
