+++ src/renderer/components/templates/index.ts
@@ -1,3 +1,2 @@
-export { default as CreateProxy } from "./CreateProxy";
-export { default as EditProxy } from "./EditProxy";
 export { default as ProxiesList } from "./ProxiesList";
+export { default as ProxyForm } from "./ProxyForm";
